requirements.txt
pytesseract
opencv-python
Pillow
google-generativeai
pdf2image
pandas
openpyxl
streamlit
langdetect
llm.gemini_processor
